for the demo, do not keep old revisions around
Deployment has its own built-in rollback mechanism, the "replicaSet"
history, which is great, but we are using Helm and Git, which both carry
numerous better opportunities for us to provide rollback capabilities to
our devs.

With revisionHistoryLimit: 0 we tell Deployments to clean replicasets
up automatically when deploy succeeds, as they are no longer needed.